# docs/swagger.yaml
basePath: /
definitions:
  github_com_fernandovalenzuela_api-home-pay_internal_models.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_fernandovalenzuela_api-home-pay_internal_models.Company:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      website_url:
        type: string
    type: object
...
  github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API REST para gestión de gastos e ingresos del hogar
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: api-home-pay API
  version: "1.0"
paths:
  /api/categories:
    get:
      consumes:
      - application/json
      description: Get all categories for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Category'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new expense category
      parameters:
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - categories
  /api/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Get a specific category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - categories
  /api/companies:
    get:
      consumes:
      - application/json
      description: Get all companies for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Company'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all companies
      tags:
      - companies
    post:
      consumes:
      - application/json
      description: Create a new company (service provider)
      parameters:
      - description: Company data
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Company'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create company
      tags:
      - companies
  /api/companies/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a company by ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete company
      tags:
      - companies
    get:
      consumes:
      - application/json
      description: Get a specific company by ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get company by ID
      tags:
      - companies
    put:
      consumes:
      - application/json
      description: Update an existing company
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      - description: Company data
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Company'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update company
      tags:
      - companies
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"